reorganized services to differentiate services between desktop and service projects easily
