+++ inventarios-desktop/src/main/java/inventarios/service/restclient/BillingService.java
@@ -1,4 +1,4 @@
-package inventarios.service;
+package inventarios.service.restclient;
 
 import inventarios.to.BillingDetails;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -6,10 +6,10 @@ import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpServerErrorException;
 import org.springframework.web.client.RestTemplate;
 
 import java.util.List;
-import org.springframework.web.client.HttpServerErrorException;
 
 @Service
 public class BillingService {
