+++ inventarios-desktop/src/main/java/inventarios/service/restclient/EmployeeService.java
@@ -14,11 +14,9 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package inventarios.service;
+package inventarios.service.restclient;
 
-import inventarios.to.OrderDetail;
-import inventarios.to.Product;
-import java.util.List;
+import inventarios.to.EmployeeDetail;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpMethod;
@@ -26,27 +24,29 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestTemplate;
 
-/**
- *
- * @author Ruslan López Carro <scherzo16 at gmail.com>
- */
+import java.util.List;
+
 @Service
-public class OrderService {
-    @Autowired
+public class EmployeeService {
     private RestTemplate restTemplate;
-    
-    public List<OrderDetail> findAll() {
-        ResponseEntity<List<OrderDetail>> response = restTemplate.exchange(
-                "http://localhost:8080/orders/",
+
+    @Autowired
+    public EmployeeService(RestTemplate restTemplate) {
+        this.restTemplate = restTemplate;
+    }
+
+    public void save(EmployeeDetail employee) {
+        restTemplate.getForEntity("http://localhost:8080/employees", EmployeeDetail.class, employee);
+    }
+
+    public List<EmployeeDetail> findAll() {        
+        ResponseEntity<List<EmployeeDetail>> response = restTemplate.exchange(
+                "http://localhost:8080/employees/",
                 HttpMethod.GET,
                 null,
-                new ParameterizedTypeReference<List<OrderDetail>>() {
+                new ParameterizedTypeReference<List<EmployeeDetail>>() {
         });
-        List<OrderDetail> employees = response.getBody();
+        List<EmployeeDetail> employees = response.getBody();
         return employees;
-    }
-    
-    public void save(OrderDetail orderDetail) {
-        restTemplate.getForEntity("http://localhost:8080/orders", OrderDetail.class, orderDetail);
     }
 }
