+++ inventarios-service/src/main/java/inventarios/service/BillingService.java
@@ -2,9 +2,10 @@ package inventarios.service;
 
 import inventarios.repository.BillingRepository;
 import inventarios.to.BillingDetails;
-import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 public class BillingService {
