+++ inventarios-service/src/test/java/inventarios/service/BillingServiceTest.java
@@ -18,16 +18,17 @@ package inventarios.service;
 
 import inventarios.repository.BillingRepository;
 import inventarios.to.BillingDetails;
-import inventarios.to.EmployeeDetail;
-import java.util.Collections;
-import static org.junit.Assert.assertEquals;
-import org.junit.Test;
 import org.junit.Before;
+import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import static org.mockito.Mockito.when;
 import org.mockito.runners.MockitoJUnitRunner;
+
+import java.util.Collections;
+
+import static org.junit.Assert.assertEquals;
+import static org.mockito.Mockito.when;
 
 /**
  *
