+++ inventarios-service/src/test/java/inventarios/service/EmployeeServiceTest.java
@@ -18,16 +18,18 @@ package inventarios.service;
 
 import inventarios.repository.EmployeeRepository;
 import inventarios.to.EmployeeDetail;
-import java.util.Collections;
-import static org.junit.Assert.assertEquals;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
-import static org.mockito.Matchers.eq;
 import org.mockito.Mock;
-import static org.mockito.Mockito.when;
 import org.mockito.runners.MockitoJUnitRunner;
+
+import java.util.Collections;
+
+import static org.junit.Assert.assertEquals;
+import static org.mockito.Matchers.eq;
+import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
 public class EmployeeServiceTest {
