+++ inventarios-service/src/test/java/inventarios/service/OrderServiceTest.java
@@ -17,6 +17,7 @@
 package inventarios.service;
 
 import inventarios.repository.OrderRepository;
+import inventarios.service.restclient.OrderService;
 import inventarios.to.EmployeeDetail;
 import inventarios.to.OrderDetail;
 import inventarios.to.Product;
