+++ inventarios-service/src/test/java/inventarios/service/ProductServiceTest.java
@@ -17,6 +17,7 @@
 package inventarios.service;
 
 import inventarios.repository.ProductRepository;
+import inventarios.service.restclient.ProductService;
 import inventarios.to.Product;
 import inventarios.to.ProductCharacteristic;
 import org.junit.Before;
