+++ inventarios-service/src/test/java/inventarios/service/ProviderServiceTest.java
@@ -17,6 +17,7 @@
 package inventarios.service;
 
 import inventarios.repository.ProviderRepository;
+import inventarios.service.restclient.ProviderService;
 import inventarios.to.Provider;
 import org.junit.Before;
 import org.junit.Test;
